+++ sqlite3_web/lib/sqlite3_web.dart
@@ -1,3 +1,4 @@
+export 'src/channel.dart' show ChannelClosedException;
 export 'src/types.dart';
 export 'src/database.dart' hide lockTokenFromId, lockTokenToId;
 export 'src/worker_connector.dart';